[feat] Add rooms
I made a Room and it's like a lobby actually.
When user hosts the game he goes to room. When user joins to host he also goes to the room. Room represents a lobby for 2-4 players and displays their  readiness. All data is synchronized between players in this room. And only the host can start the game.
Also added a small feature for some time - unique background color for each player in room.