[feat] Add usernames
Added usernames. Now on the first game load player can set the name before the connection to the room. Name should be not equal to null and be less than 11 chars. Name of the user is displayed in the room. So everyone in the room can see who is ready/not ready. Also user can reset his name by clicking special button.